+++ libs/firebase-utils/src/lib/firestoreMigrations/index.ts
@@ -41,7 +41,6 @@ import * as v0040 from './0040';
 import * as v0041 from './0041';
 import * as v0042 from './0042';
 import * as v0043 from './0043';
-import * as v0044 from './0044';
 
 import { Firestore, Storage } from '../types';
 
@@ -96,8 +95,7 @@ export const MIGRATIONS = {
   40: v0040,
   41: v0041,
   42: v0042,
-  43: v0043,
-  44: v0044
+  43: v0043
 };
 
 export const VERSIONS_NUMBERS = Object.keys(MIGRATIONS).map(s => parseInt(s, 10));
